fix(story): add deticated component for ai summary
